Moved slCamera and removed imports
